refactored fullname in auth provider
